Modularised the original screen.py into a textStarSerialLCD module, and an example of its use in screen.py  baud_rate, poll tick, refresh ticks and the control characters can now be set via kwargs.
