+++ screen.py
@@ -7,13 +7,13 @@ screen - Manages the Textstar 16x2 4 button display
 
 Read the blog entry at http://jeremyblythe.blogspot.com for more information
 '''
-import serial
 import datetime
 import time
 import subprocess
 import twitter
 import urllib2
 import json
+import textStarSerialLCD
 
 CLEAR = chr(12)
 ESC = chr(254)
@@ -22,71 +22,10 @@ BLOCK = chr(154)
 POLL_TICKS = 15
 REFRESH_TICKS = 300
 
-class Display():
-    """ Manages the 16x2 4 button display:
-        on_tick called every 0.1 seconds as part of the main loop after the button read
-        on_poll called every 1.5 seconds
-        on_page called when a new page has been selected
-        on_refresh called every 30 seconds"""
-    def __init__(self,on_page=None,on_poll=None,on_tick=None,on_refresh=None):
-        # Start the serial port
-        self.ser = serial.Serial('/dev/ttyAMA0',9600,timeout=0.1)
-        # Callbacks
-        self.on_page = on_page
-        self.on_poll = on_poll
-        self.on_tick = on_tick
-        self.on_refresh = on_refresh
-
-        self.page = 'a'
-        self.poll = POLL_TICKS
-        self.refresh = REFRESH_TICKS
-
-    def position_cursor(self,line,column):
-        self.ser.write(ESC+'P'+chr(line)+chr(column))
-    
-    def scroll_down(self):
-        self.ser.write(ESC+'O'+chr(0))
-    
-    def window_home(self):
-        self.ser.write(ESC+'G'+chr(1))
-
-    def clear(self):
-        self.ser.write(CLEAR)
-
-    def run(self):
-        #show initial page
-        display.ser.write('  Starting....  ')
-        if self.on_page != None:
-            self.on_page()
-        #main loop
-        while True:
-            key = str(self.ser.read(1))
-            if key != '' and key in 'abcd':
-                self.page = key
-                self.refresh = REFRESH_TICKS
-                self.poll = POLL_TICKS
-                if self.on_page != None:
-                    self.on_page()
-            else:
-                self.refresh-=1
-                if self.refresh == 0:
-                    self.refresh = REFRESH_TICKS
-                    if self.on_refresh != None:
-                        self.on_refresh()
-                        
-                self.poll-=1
-                if self.poll == 0:
-                    self.poll = POLL_TICKS
-                    if self.on_poll != None:
-                        self.on_poll()
-                        
-                if self.on_tick != None:
-                    self.on_tick()
-
-
 display = None
 # Start twitter
 twitter_api = twitter.Api()
+
 
 def write_datetime():
     display.position_cursor(1, 1)
@@ -157,6 +96,6 @@ def on_refresh():
     elif display.page == 'd':
         write_ip_addresses()
 
-display = Display(on_page, on_poll, on_tick, on_refresh)            
+display = textStarSerialLCD.Display(on_page, on_poll, on_tick, on_refresh)            
 display.run()
